+++ src/components/StatusBadge/StatusBadge.tsx
@@ -1,8 +1,16 @@
 import { Box, chakra, ChakraComponent, useStyleConfig } from "@chakra-ui/react";
 import React, { forwardRef } from "react";
 
+export const statusBadgeFontSizeArray = [
+  "desktop.body.body1",
+  "desktop.body.body2",
+  "desktop.caption",
+] as const;
+export type StatusBadgeFontSizes = typeof statusBadgeFontSizeArray[number];
+
 export const statusBadgeLevelArray = ["low", "medium", "high"] as const;
 export type StatusBadgeLevels = typeof statusBadgeLevelArray[number];
+
 export const statusBadgeTypeArray = [
   "informative",
   "negative",
@@ -34,17 +42,20 @@ export interface StatusBadgeProps {
  */
 export const StatusBadge: ChakraComponent<
   React.ForwardRefExoticComponent<
-    StatusBadgeProps & {
-      children?: React.ReactNode;
-    } & React.RefAttributes<HTMLDivElement>
+    React.PropsWithChildren<StatusBadgeProps> &
+      React.RefAttributes<HTMLDivElement>
   >,
-  StatusBadgeProps
+  React.PropsWithChildren<StatusBadgeProps>
 > = chakra(
   forwardRef<HTMLDivElement, React.PropsWithChildren<StatusBadgeProps>>(
     (props, ref?) => {
       const { children, className, id, level, type, ...rest } = props;
+      const labelFontSize = rest["fontSize"] || "desktop.body.body2";
       const finalVariant = level ? level : type ? type : "low";
-      const styles = useStyleConfig("StatusBadge", { variant: finalVariant });
+      const styles = useStyleConfig("StatusBadge", {
+        labelFontSize,
+        variant: finalVariant,
+      });
 
       if (!children) {
         console.warn("NYPL Reservoir StatusBadge: No children were passed.");